+++ app/src/components/index.tsx
@@ -55,3 +55,4 @@ export * from "./dialog";
 export * from "./tooltip";
 export * from "./progressCircle";
 export * from "./breadcrumbs";
+export * from "./progress";