+++ app/src/components/progress/index.tsx
@@ -1,1 +1,2 @@
 export { ProgressCircle } from "./ProgressCircle";
+export { ProgressBar } from "./ProgressBar";